Create IEEE123testfederForAR
The application of the algorithm is defined by an simple search for solution automATIC RECONFIGURATION USING  for  the IEEE 123 test feeder. The basic solution is the find the shortest path from the source(transformer) to all recloser (8) in the grid. A post processing method was devised to aid in a reconfiguration using multi-agent systems approch also devised by me.